Fix drag and drop for Xcb
Make the xcb plugin work in native coordinates, using platform windows as
much as possible. As part of this, replace QMouseEvent in QBasicDrag with
just a QPoint. This is private API, so we can change this without warning.

In addition to xcb, QBasicDrag is used in QSimpleDrag and for Wayland.
QSimpleDrag does not yet have a fix for highDpi, but it should continue
working in the non-scaled case. Wayland (and any other module which uses
QBasicDrag) will need a compile fix.

Also fix bug in QWindowSystemInterface: handleDrag()/handleDrop()
take positions in window local coordinates, not global.

Change-Id: I86543e4f52a7b3ba1efeac815cf89bbd97c0a0a2
Reviewed-by: Morten Johan Sørvig <morten.sorvig@theqtcompany.com>