Add a qt_version_tag symbol to QtCore that uses ELF versions
This symbol will exist with different ELF versions, allowing us to know
which version of Qt a given library or application was linked
against. When this symbol gets used, automatic packaging tools will
extract the dependency. The trick here is that the application uses
qt_version_tag without knowing which version is current, but the linker
resolves it to the current version and records that.

For example, if this were used with Qt 5.5, RPM's find-requires on an
application using QtCore would print:

  libQt5Core.so.5()(64bit)
  libQt5Core.so.5(Qt_5.5)(64bit)
  libQt5Core.so.5(Qt_5)(64bit)

Whereas find-provides on QtCore would print:

  libQt5Core.so.5()(64bit)
  libQt5Core.so.5(libQt5Core.so.5)(64bit)
  libQt5Core.so.5(Qt_5.0)(64bit)
  libQt5Core.so.5(Qt_5.1)(64bit)
  libQt5Core.so.5(Qt_5.2)(64bit)
  libQt5Core.so.5(Qt_5.3)(64bit)
  libQt5Core.so.5(Qt_5.4)(64bit)
  libQt5Core.so.5(Qt_5.5)(64bit)
  libQt5Core.so.5(Qt_5)(64bit)

Therefore, automatic dependency resolution would have the information it
needs to conclude that the application requires Qt >= 5.5.

Change-Id: I049a653beeb5454c9539ffff13e3fec9aeb50197
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>