Unexport QSslPreSharedKeyAuthenticator
Exporting value classes (as opposed to just their non-inline
methods) creates subtle binary incompatibility problems.

In this case, between C++11 and C++98 builds because of the
move assignment operator.

Even though it's not a problem in practice, so far, for some
types of classes this issue ie real (QVector, say), so it's
best to avoid exporting what we don't need to export.

Change-Id: Ifca6aaedcbfa79ca35e651de7630e69c3b266fe3
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>