Scale logicalDpi when turning on screen scaling
Assume that the logical DPI for the primary screen is sane, and
keep font sizes constant as screens are scaled. The global scale
factor will act as a zoom level, and will cause all fonts to be bigger.

Note that since we do not change logicalDpi after application startup,
the manual test sliders will not match what happens
with auto scaling. We may want to fix that...

Change-Id: I5a3daa57c2dacf0158836492d31573723e49399a
Reviewed-by: Morten Johan Sørvig <morten.sorvig@theqtcompany.com>