CoreText: Respect Qt::AA_Use96Dpi flag when creating CT fonts
The pixel size passed into QCoreTextFontDatabase::fontEngine() has already
been scaled to take the destination DPI into account. In the normal case
the destination DPI is the same as the screen's logical DPI, 72, and we
end up with the pixel size being equal to the point size. But when the
Qt::AA_Use96Dpi flag has been set, e.g. when running tests, the pixel
size is larger than the point size, and we need to pass this information
to CoreText, otherwise we end up with clipped text rendering. This was
evident when looking at the lancelot baselines for OS X, which are
produced using QTestLib, and hence run with Qt::AA_Use96Dpi set.

Change-Id: Ib0f720bb318b2dc437172a6b712675e3dfe3f7ac
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@theqtcompany.com>